Ganache now reverts state between tests
