+++ package.json
@@ -15,6 +15,7 @@
   "license": "MIT",
   "devDependencies": {
     "@types/jest": "^26.0.24",
+    "ganache-core": "^2.13.2",
     "jest": "^27.0.6",
     "ts-jest": "^27.0.4",
     "typescript": "^4.3.5"
